update nodejs to 5.4.1

+++ net-libs/nodejs/nodejs-5.4.1.ebuild
@@ -4,33 +4,39 @@
 
 EAPI=5
 
-# has known failures. sigh.
-RESTRICT="test"
-
 PYTHON_COMPAT=( python2_7 )
-PYTHON_REQ_USE="threads"
 
-inherit pax-utils python-single-r1 toolchain-funcs
+inherit flag-o-matic pax-utils python-single-r1 toolchain-funcs
 
 DESCRIPTION="Evented IO for V8 Javascript"
 HOMEPAGE="http://nodejs.org/"
-SRC_URI="http://nodejs.org/dist/v${PV}/node-v${PV}.tar.gz"
+SRC_URI="http://nodejs.org/dist/v${PV}/node-v${PV}.tar.xz"
 
 LICENSE="Apache-1.1 Apache-2.0 BSD BSD-2 MIT"
 SLOT="0"
 KEYWORDS="amd64"
-IUSE="debug icu +npm +snapshot +ssl"
+IUSE="debug icu +npm snapshot +ssl"
 
-RDEPEND="icu? ( dev-libs/icu )
+RDEPEND="icu? ( >=dev-libs/icu-55:= )
 	${PYTHON_DEPS}
-	ssl? ( dev-libs/openssl:0=[-bindist] )
-	>=net-libs/http-parser-2.3
-	>=dev-libs/libuv-1.4.2"
+	>=net-libs/http-parser-2.5:=
+	>=dev-libs/libuv-1.6.1:=
+	>=dev-libs/openssl-1.0.2d:0=[-bindist]
+	sys-libs/zlib
+"
 DEPEND="${RDEPEND}
 	!!net-libs/iojs"
 
 S="${WORKDIR}/node-v${PV}"
 REQUIRED_USE="${PYTHON_REQUIRED_USE}"
+
+pkg_pretend() {
+	if [[ ${MERGE_TYPE} != "binary" ]] ; then
+		if ! test-flag-CXX -std=c++11 ; then
+			die "Your compiler doesn't support C++11. Use GCC 4.8, Clang 3.3 or newer."
+		fi
+	fi
+}
 
 src_prepare() {
 	tc-export CC CXX PKG_CONFIG
@@ -38,7 +44,7 @@ src_prepare() {
 	export BUILDTYPE=Release
 
 	# fix compilation on Darwin
-	# http://code.google.com/p/gyp/issues/detail?id=260
+	# https://code.google.com/p/gyp/issues/detail?id=260
 	sed -i -e "/append('-arch/d" tools/gyp/pylib/gyp/xcode_emulation.py || die
 
 	# make sure we use python2.* while using gyp
@@ -54,36 +60,59 @@ src_prepare() {
 	sed -i -e "s/'lib'/'${LIBDIR}'/" lib/module.js || die
 	sed -i -e "s|\"lib\"|\"${LIBDIR}\"|" deps/npm/lib/npm.js || die
 
+	# Avoid a test that I've only been able to reproduce from emerge. It doesnt
+	# seem sandbox related either (invoking it from a sandbox works fine).
+	# The issue is that no stdin handle is openened when asked for one.
+	# It doesn't really belong upstream , so it'll just be removed until someone
+	# with more gentoo-knowledge than me (jbergstroem) figures it out.
+	rm test/parallel/test-stdout-close-unref.js || die
+	# AssertionError: 1 == 2 (on line 97)
+	rm test/parallel/test-cluster-disconnect.js || die
+	# AssertionError: Client never errored
+	rm test/parallel/test-tls-hello-parser-failure.js || die
+	# --- TIMEOUT ---
+	rm test/parallel/test-child-process-fork-net.js \
+		test/parallel/test-child-process-fork-net2.js \
+		test/parallel/test-child-process-recv-handle.js \
+		test/parallel/test-cluster-dgram-1.js \
+		test/parallel/test-cluster-send-deadlock.js \
+		test/parallel/test-cluster-shared-handle-bind-error.js \
+		test/parallel/test-dgram-exclusive-implicit-bind.js \
+		test/parallel/test-tls-ticket-cluster.js || die
+
 	# debug builds. change install path, remove optimisations and override buildtype
 	if use debug; then
 		sed -i -e "s|out/Release/|out/Debug/|g" tools/install.py || die
 		BUILDTYPE=Debug
 	fi
+
+	epatch_user
 }
 
 src_configure() {
-	local myconf=()
 	local myarch=""
-	use debug && myconf+=( --debug )
-	use icu && myconf+=( --with-intl=system-icu )
+	local myconf+=( --shared-openssl --shared-libuv --shared-http-parser --shared-zlib )
 	use npm || myconf+=( --without-npm )
-	use snapshot || myconf+=( --without-snapshot )
+	use icu && myconf+=( --with-intl=system-icu )
+	use snapshot && myconf+=( --with-snapshot )
 	use ssl || myconf+=( --without-ssl )
+	use debug && myconf+=( --debug )
 
 	case ${ABI} in
 		x86) myarch="ia32";;
 		amd64) myarch="x64";;
+		x32) myarch="x32";;
 		arm) myarch="arm";;
+		arm64) myarch="arm64";;
 		*) die "Unrecognized ARCH ${ARCH}";;
 	esac
 
+	GYP_DEFINES="linux_use_gold_flags=0
+		linux_use_bundled_binutils=0
+		linux_use_bundled_gold=0" \
 	"${PYTHON}" configure \
 		--prefix="${EPREFIX}"/usr \
 		--dest-cpu=${myarch} \
-		--shared-openssl \
-		--shared-libuv \
-		--shared-http-parser \
-		--shared-zlib \
 		--without-dtrace \
 		"${myconf[@]}" || die
 }
@@ -97,9 +126,11 @@ src_compile() {
 src_install() {
 	local LIBDIR="${ED}/usr/$(get_libdir)"
 	emake install DESTDIR="${ED}" PREFIX=/usr
-	use npm && dodoc -r "${LIBDIR}"/node_modules/npm/html
-	rm -rf "${LIBDIR}"/node_modules/npm/{doc,html} || die
-	find "${LIBDIR}"/node_modules -type f -name "LICENSE*" -or -name "LICENCE*" -delete
+	if use npm; then
+		dodoc -r "${LIBDIR}"/node_modules/npm/html
+		rm -rf "${LIBDIR}"/node_modules/npm/{doc,html} || die
+		find "${LIBDIR}"/node_modules -type f -name "LICENSE*" -or -name "LICENCE*" -delete || die
+	fi
 
 	# set up a symlink structure that npm expects..
 	dodir /usr/include/node/deps/{v8,uv}
@@ -112,8 +143,9 @@ src_install() {
 }
 
 src_test() {
+	out/${BUILDTYPE}/cctest || die
 	declare -xl TESTTYPE="${BUILDTYPE}"
-	"${PYTHON}" tools/test.py --mode=${TESTTYPE} -J message simple || die
+	"${PYTHON}" tools/test.py --mode=${TESTTYPE} -J message parallel sequential || die
 }
 
 pkg_postinst() {
